# backend/src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/binary/full/": {
            "post": {
                "description": "Uploads a file and processes it for a task",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "plain/text"
                ],
                "summary": "Upload a file for a full task",
                "parameters": [
                    {
                        "type": "file",
                        "description": "File to upload",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/structs.FileTaskResponse"
                        }
                    },
                    "400": {
                        "description": "Error retrieving the file",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Error opening or reading the file",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/binary/short": {
            "post": {
                "description": "Uploads a text and processes it for a task",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Upload a text for a short task",
                "parameters": [
                    {
                        "description": "Text for a task",
                        "name": "text",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/structs.TextTaskRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/structs.TextTaskResponse"
                        }
                    },
                    "400": {
                        "description": "Error retrieving the file",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Error opening or reading the file",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "messages.MessageResult": {
            "type": "object",
            "properties": {
                "messageText": {
                    "type": "string"
                },
                "result": {
                    "type": "integer"
                },
                "submitDate": {
                    "type": "string"
                },
                "userID": {
                    "type": "string"
                }
            }
        },
        "structs.FileTaskResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/messages.MessageResult"
                    }
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "structs.TextTaskRequest": {
            "type": "object",
            "properties": {
                "text": {
                    "type": "string"
                }
            }
        },
        "structs.TextTaskResponse": {
            "type": "object",
            "properties": {
                "result": {
                    "type": "integer"
                },
                "text": {
                    "type": "string"
                }
            }
        }
    }
}